% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{bussin}
\alias{bussin}
\title{Bussin analysis}
\usage{
bussin(.data)
}
\arguments{
\item{.data}{A data frame or tibble}
}
\value{
The data frame (for piping)
}
\description{
Chain multiple operations when your analysis is absolutely bussin
}
