% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{clout_chase}
\alias{clout_chase}
\title{right_join()	clout_chase()}
\usage{
clout_chase(x, y, ...)
}
\arguments{
\item{x}{A data frame or tibble (main squad)}

\item{y}{A data frame or tibble (joining squad)}

\item{by}{Columns to join by}
}
\value{
A merged data frame
}
\description{
Merge, but the other chums call the shots on who stays (right table is main squad)
}
