% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{disband}
\alias{disband}
\title{Disband the squad}
\usage{
disband(.data)
}
\arguments{
\item{.data}{A grouped data frame}
}
\value{
An ungrouped data frame
}
\description{
Remove grouping (ungroup equivalent)
}
\examples{
mtcars |> squad_up(cyl) |> disband()
}
