% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{fam}
\alias{fam}
\title{bind_rows()  fam()}
\usage{
fam(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Data frames to combine. Each argument can either be a data frame,
a list that could be a data frame, or a list of data frames.}
}
\value{
A data frame the same type as the first element of \code{...}.
}
\description{
Append some homies to a data frame.
}
\examples{
df1 <- data.frame(id = "bro", x = 1:2, y = letters[1:2])
df2 <- data.frame(id = "sis", x = 4:5, z = 1:2)
fam(df1,df2)
}
