% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{fit_check}
\alias{fit_check}
\title{Fit check}
\value{
x original x is (invisibly) returned, allowing \code{fit_check()} to be
used within a data pipeline.
}
\description{
Don't get shooketh, spill the tea on your data frame
}
