% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{ghost}
\alias{ghost}
\title{anti_join()	ghost()}
\usage{
ghost(x, y, ...)
}
\arguments{
\item{x}{A data frame or tibble}

\item{y}{A data frame or tibble}
}
\value{
A data frame with non-matching rows
}
\description{
Forget about those who don't vibe with you (remove non-matching rows)
}
