% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{its_giving}
\alias{its_giving}
\title{Its giving... count occurrences}
\usage{
its_giving(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Variables to count by}
}
\value{
A data frame with counts
}
\description{
Count observations by group (count equivalent)
}
\examples{
mtcars |> its_giving(cyl, gear)
}
