% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{link_up}
\alias{link_up}
\title{left_join()	link_up()}
\usage{
link_up(x, y, ...)
}
\arguments{
\item{x}{A data frame or tibble (main squad)}

\item{y}{A data frame or tibble (joining squad)}

\item{by}{Columns to join by}
}
\value{
A merged data frame
}
\description{
Merge datasets but keep only your day-ones (left table stays main squad)
}
