% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{lowkey}
\alias{lowkey}
\title{Lowkey rename columns}
\usage{
lowkey(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Name-value pairs (new_name = old_name)}
}
\value{
A data frame with renamed columns
}
\description{
Change column names on the down-low (rename equivalent)
}
\examples{
mtcars |> lowkey(miles_per_gallon = mpg, horse_power = hp)
}
