% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{main_character}
\alias{main_character}
\title{Main character energy - pull a column}
\usage{
main_character(.data, var)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{var}{Column to extract}
}
\value{
A vector
}
\description{
Extract a column as a vector (pull equivalent)
}
\examples{
mtcars |> main_character(mpg)
}
