% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{mutuals_only}
\alias{mutuals_only}
\title{inner_join()	mutuals_only()}
\usage{
mutuals_only(x, y, ...)
}
\arguments{
\item{x}{A data frame or tibble}

\item{y}{A data frame or tibble}

\item{by}{Columns to join by}
}
\value{
A merged data frame
}
\description{
Merge, but only keep rows where both tables are mutually following each other
}
