% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{no_cap}
\alias{no_cap}
\title{No cap summary stats}
\usage{
no_cap(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Name-value pairs of summary expressions}
}
\value{
A data frame with summary statistics
}
\description{
Get the real stats, no cap (summarise equivalent)
}
\examples{
mtcars |> no_cap(avg_mpg = mean(mpg), max_hp = max(hp))
}
