% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{only_the_reals}
\alias{only_the_reals}
\title{semi_join()	only_the_reals()}
\usage{
only_the_reals(x, y, ...)
}
\arguments{
\item{x}{A data frame or tibble}

\item{y}{A data frame or tibble}
}
\value{
A data frame with matching rows
}
\description{
Just keep the rows that vibe with both tables (matching rows only)
}
