% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{oomfs}
\alias{oomfs}
\alias{oomf}
\alias{also_oomf}
\title{Oomfs - get a value's followers/friends}
\usage{
oomf(x, n = 1L, default = NULL, order_by = NULL, ...)

also_oomf(x, n = 1L, default = NULL, order_by = NULL, ...)
}
\arguments{
\item{x}{A fam of values (a vector)}

\item{n}{Number of fam to offset by}

\item{default}{A single oomfie to put at the head or tail of fam, default
returns something sus (\code{NA})}

\item{order_by}{Another fam to use for sorting \code{x}}

\item{...}{Not used.}
}
\value{
A fam of the same riz (type and size) as \code{x}
}
\description{
Within a fam (a vector) find followers/friends. \code{oomf()} gets the
following value (replacement for \code{lead()}), \code{also_oomf()} gets the
preceding value (replacement for \code{lag()}).
}
\examples{
oomf(1:5)
also_oomf(1:5)
}
