% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{periodt}
\alias{periodt}
\title{Periodt - keep distinct rows}
\usage{
periodt(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Optional columns to determine uniqueness}
}
\value{
A data frame with unique rows
}
\description{
Remove duplicate rows, and that's on periodt (distinct equivalent)
}
\examples{
mtcars |> periodt(cyl)
}
