% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{plot_twist}
\alias{plot_twist}
\title{case_when()	plot_twist()}
\usage{
plot_twist(...)
}
\arguments{
\item{...}{A sequence of two-sided formulas. On the left: a logical condition (the vibe check). On the right: the value to assign when the vibe is true.}
}
\value{
A vector whose length matches the input; each element gets the first matching value or \code{NA} if none match.
}
\description{
Switch up the plot depending on vibes (case_when equivalent)
}
\examples{
mtcars |> glow_up(vibe = plot_twist(mpg > 20 ~ "bussin", TRUE ~ "mid"))
}
