% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{rent_free}
\alias{rent_free}
\title{Living rent free in your head (and data)}
\usage{
rent_free(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Columns to move}
}
\value{
A data frame with rearranged columns
}
\description{
Move columns to the front or specific positions because they matter (relocate equivalent)
}
\examples{
# Move 'hp' to the front because it's living rent free
mtcars |> rent_free(hp)
}
