% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{say_less}
\alias{say_less}
\title{Say less... return number of unique occurrences}
\usage{
say_less(...)
}
\arguments{
\item{...}{a vector or multiple vectors}
}
\value{
A single number
}
\description{
Number of unique occurrences in a vector or vectors (count equivalent)
}
\examples{
n_distinct(mtcars$cyl)
}
