% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{send_it}
\alias{send_it}
\title{Send it - take the first n rows}
\usage{
send_it(.data, n = 6)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{n}{Number of rows to keep}
}
\value{
A data frame with first n rows
}
\description{
Get the top rows (head/slice_head equivalent)
}
\examples{
mtcars |> send_it(5)
}
