% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{situationship}
\alias{situationship}
\title{bind_cols()	situationship()}
\usage{
situationship(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Data frames to combine. Each argument can either be a data frame,
a list that could be a data frame, or a list of data frames.}
}
\value{
A data frame the same type as the first element of \code{...}.
}
\description{
Combine data frames that are not quite a couple but a bit more than friends.
}
\examples{
mtcars[1:3,] |> situationship(him = 1:3, her = 3:1)
}
