% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{six_seven}
\alias{six_seven}
\title{SIX-SEVEN!!!!!!!!!}
\usage{
six_seven(n)
}
\arguments{
\item{n}{HOW MANY TIMES ARE YOU YELLING SIX-SEVEN?!!!!!!}
}
\description{
SIX-SEVEN! SIX-SEVEN! SIX-SEVEN! SIX-SEVEN! SIX-SEVEN! SIX-SEVEN! SIX-SEVEN!
}
