% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{slay}
\alias{slay}
\title{Sort by slay factor}
\usage{
slay(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Variables to sort by}
}
\value{
A sorted data frame
}
\description{
Arrange rows by columns (arrange equivalent)
}
\examples{
mtcars |> slay(desc(mpg), hp)
}
