% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{squad_up}
\alias{squad_up}
\title{Squad up your data}
\usage{
squad_up(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Variables to group by}
}
\value{
A grouped data frame
}
\description{
Group by categories to analyze squads (group_by equivalent)
}
\examples{
mtcars |> squad_up(cyl) |> no_cap(avg_mpg = mean(mpg))
}
