% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{thick_of_it}
\alias{thick_of_it}
\title{Joining in The Thick Of It}
\usage{
thick_of_it(x, y, type, ...)
}
\arguments{
\item{x, y}{Twinnin data frames, tibbles, or lazy data frames (where have they been?)}

\item{type}{The type of join you're using - can be "inner", "left", "right", or "full"}

\item{...}{Additional parameters to be passed into the join functions}
}
\description{
Join everything from the screen, to the ring, to the pen, to the king...
}
