% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{vibe_check}
\alias{vibe_check}
\title{Vibe check your columns}
\usage{
vibe_check(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Columns to keep}
}
\value{
A data frame with selected columns
}
\description{
Keep only the columns that pass the vibe check (select equivalent)
}
\examples{
mtcars |> vibe_check(mpg, cyl, hp)
}
