% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{yeet}
\alias{yeet}
\title{Yeet rows from your data}
\usage{
yeet(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Logical predicates to keep rows}
}
\value{
A data frame with yeeted rows removed
}
\description{
Remove rows that don't pass the vibe check (filter equivalent)
}
\examples{
mtcars |> yeet(mpg > 20, cyl == 4)
}
