% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genzplyr.R
\name{six_seven}
\alias{six_seven}
\title{Six seven - random sample}
\usage{
six_seven(.data, ..., n, prop, by = NULL, weight_by = NULL, replace = FALSE)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Additional arguments passed to slice_sample()}

\item{n}{Number of rows to sample}

\item{prop}{Proportion of rows to sample (alternative to n)}

\item{by}{Variables to group by}

\item{weight_by}{Sampling weights}

\item{replace}{Sample with replacement}
}
\value{
A data frame with randomly sampled rows
}
\description{
Take a random sample of rows (slice_sample equivalent)
}
\examples{
mtcars |> six_seven(n = 5)
mtcars |> six_seven(prop = 0.25)
}
